# openapi.yaml
openapi: 3.0.3
info:
  title: ms-pedidos API
  description: |
    Microservicio de gestión de pedidos.
    **Regla clave:** el campo `total` siempre se calcula en el backend:
    `total = cantidad × precioUnitario`. El cliente no puede enviarlo.
    La eliminación es lógica (estado → CANCELADO).
  version: 1.0.0
  contact:
    name: Miguel Sánchez
    email: sanchezsanchezmiguelivan@gmail.com

servers:
  - url: http://localhost:8081
    description: Local
  - url: https://ms-pedidos.onrender.com
    description: Render (producción)

tags:
  - name: Pedidos
    description: Operaciones de gestión de pedidos

paths:
  /api/pedidos:
    post:
      tags: [Pedidos]
      summary: Crear pedido
      description: |
        Registra un nuevo pedido. El total se calcula automáticamente:
        `total = cantidad × precioUnitario`. Estado inicial: REGISTRADO.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoRequest'
            example:
              cliente: Juan Pérez
              correoCliente: juan@email.com
              productoId: 1
              nombreProducto: Laptop Lenovo
              cantidad: 2
              precioUnitario: 3500.00
      responses:
        '201':
          description: Pedido creado con total calculado automáticamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
              example:
                id: 1
                cliente: Juan Pérez
                correoCliente: juan@email.com
                productoId: 1
                nombreProducto: Laptop Lenovo
                cantidad: 2
                precioUnitario: 3500.00
                total: 7000.00
                estado: REGISTRADO
                fechaPedido: "2026-05-17T15:48:18.675"
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Pedidos]
      summary: Listar pedidos
      description: Retorna todos los pedidos registrados.
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PedidoResponse'

  /api/pedidos/{id}:
    get:
      tags: [Pedidos]
      summary: Buscar pedido por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                mensaje: Pedido no encontrado
                detalle: No existe un pedido con el ID 5
                fecha: "2026-05-17T15:42:46.980"

    delete:
      tags: [Pedidos]
      summary: Eliminar pedido (lógico)
      description: Eliminación lógica — cambia el estado a `CANCELADO`. No borra físicamente el registro.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Pedido cancelado (sin cuerpo de respuesta)
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/pedidos/{id}/estado:
    patch:
      tags: [Pedidos]
      summary: Actualizar estado del pedido
      description: |
        Actualiza únicamente el estado del pedido.
        Estados válidos: `REGISTRADO`, `PAGADO`, `ENVIADO`, `CANCELADO`
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstadoUpdate'
            example:
              estado: PAGADO
      responses:
        '200':
          description: Estado actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Estado inválido o vacío
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PedidoRequest:
      type: object
      required:
        - cliente
        - correoCliente
        - productoId
        - nombreProducto
        - cantidad
        - precioUnitario
      properties:
        cliente:
          type: string
          description: Nombre del cliente (obligatorio)
          example: Juan Pérez
        correoCliente:
          type: string
          format: email
          description: Correo electrónico válido del cliente
          example: juan@email.com
        productoId:
          type: integer
          format: int64
          description: ID del producto solicitado
          example: 1
        nombreProducto:
          type: string
          description: Nombre del producto
          example: Laptop Lenovo
        cantidad:
          type: integer
          description: Cantidad solicitada (debe ser mayor que cero)
          minimum: 1
          example: 2
        precioUnitario:
          type: number
          format: decimal
          description: Precio unitario del producto (debe ser mayor que cero)
          minimum: 0.01
          example: 3500.00

    EstadoUpdate:
      type: object
      required:
        - estado
      properties:
        estado:
          type: string
          description: Nuevo estado del pedido
          enum: [REGISTRADO, PAGADO, ENVIADO, CANCELADO]
          example: PAGADO

    PedidoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        cliente:
          type: string
          example: Juan Pérez
        correoCliente:
          type: string
          example: juan@email.com
        productoId:
          type: integer
          format: int64
          example: 1
        nombreProducto:
          type: string
          example: Laptop Lenovo
        cantidad:
          type: integer
          example: 2
        precioUnitario:
          type: number
          format: decimal
          example: 3500.00
        total:
          type: number
          format: decimal
          description: "Calculado en backend: cantidad × precioUnitario"
          example: 7000.00
        estado:
          type: string
          enum: [REGISTRADO, PAGADO, ENVIADO, CANCELADO]
          example: REGISTRADO
        fechaPedido:
          type: string
          format: date-time
          description: Asignada automáticamente al crear el pedido
          example: "2026-05-17T15:48:18.675"

    ErrorResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Pedido no encontrado
        detalle:
          type: string
          example: No existe un pedido con el ID 5
        fecha:
          type: string
          format: date-time
          example: "2026-05-17T15:42:46.980"